layout/xul/base/test/window_resizer.xul

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
	hidechrome="true"
	windowtype="main"
	orient="vertical"
	sizemode="normal"
	onload="setTimeout(doTest, 0)"
	persist="screenX screenY width height">
<script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
<script><![CDATA[
var is = window.opener.SimpleTest.is;

function doTest() {
  // from test_resizer.xul
  var expectX = 200;
  var expectY = 200;
  var expectXMost = 500;
  var expectYMost = 500;
  var screenScale = expectX/window.screenX;
  var root = document.documentElement;

  function testResizer(dx, dy) {
    var offset = 20;
    var scale = 5;
    // target the centre of the resizer
    var offsetX = window.innerWidth/2 + (window.innerWidth/3)*dx;
    var offsetY = window.innerHeight/2 + (window.innerHeight/3)*dy;

    for (mouseX = -1; mouseX <= 1; ++mouseX) {
      for (mouseY = -1; mouseY <= 1; ++mouseY) {
        var newExpectX = expectX;
        var newExpectXMost = expectXMost;
        var newExpectY = expectY;
        var newExpectYMost = expectYMost;
        if (dx < 0) {
          newExpectX += mouseX*scale;
        } else if (dx > 0) {
          newExpectXMost += mouseX*scale;
        }
        if (dy < 0) {
          newExpectY += mouseY*scale;
        } else if (dy > 0) {
          newExpectYMost += mouseY*scale;
        }

        synthesizeMouse(root, offsetX, offsetY, { type:"mousedown" });
        synthesizeMouse(root, offsetX + mouseX*scale, offsetY + mouseY*scale, { type:"mousemove" });
        is(window.screenX*screenScale, newExpectX,
           "Bad x for " + dx + "," + dy + " moving " + mouseX + "," + mouseY);
        is(window.screenY*screenScale, newExpectY,
           "Bad y for " + dx + "," + dy + " moving " + mouseX + "," + mouseY);
        is(window.outerWidth, newExpectXMost - newExpectX,
           "Bad width for " + dx + "," + dy + " moving " + mouseX + "," + mouseY);
        is(window.outerHeight, newExpectYMost - newExpectY,
           "Bad height for " + dx + "," + dy + " moving " + mouseX + "," + mouseY);
        // move it back before we release! Adjust for any window movement
        synthesizeMouse(root, offsetX - (newExpectX - expectX),
                              offsetY - (newExpectY - expectY), { type:"mousemove" });
        synthesizeMouse(root, offsetX, offsetY, { type:"mouseup" });
      }
    }
  }

  testResizer(-1, -1);
  testResizer(-1, 0);
  testResizer(-1, 1);
  testResizer(0, -1);
  testResizer(0, 1);
  testResizer(1, -1);
  testResizer(1, 0);
  testResizer(1, 1);
  window.close();
  window.opener.SimpleTest.finish();
}
]]></script>
	<hbox flex="1">
		<vbox flex="1">
			<resizer dir="topleft" flex="1"/>
			<resizer dir="left" flex="1"/>
			<resizer dir="bottomleft" flex="1"/>
		</vbox>
		<vbox flex="1">
			<resizer dir="top" flex="1"/>
			<hbox flex="1"/>
			<resizer dir="bottom" flex="1"/>
		</vbox>
		<vbox flex="1">
			<resizer dir="topright" flex="1"/>
			<resizer dir="right" flex="1"/>
			<resizer dir="bottomright" flex="1"/>
		</vbox>
	</hbox>
</window>
